# EVEN_K when K is divisible by BLOCK_SIZE_K