EVEN_K flag if the K is divisible by BLOCK_SIZE_K, the masking will be disabled in GEMM 

EVEN_K enabled performance:
|     M    |      N    |     K    |  E  | top_k | Time (ms) |      TFLOPS     | Bandwidth (GB/s) |
|----------|-----------|----------|-----|-------|-----------|-----------------|------------------|
|    64.0  |    256.0  |   128.0  | 8.0 |   2.0 |   0.083316 |      0.104642   |        7.625853  |
|    64.0  |   1792.0  |  1024.0  | 8.0 |   2.0 |   0.090971 |      5.381620   |      352.872503  |
|    64.0  |   7168.0  |  4096.0  | 8.0 |   2.0 |   0.261861 |     28.611148   |     1798.869851  |
|   128.0  |   7168.0  |  4096.0  | 8.0 |   2.0 |   0.269888 |     56.154659   |     1780.769457  |
|  1024.0  |   7168.0  |  4096.0  | 8.0 |   2.0 |   0.672297 |    185.206336   |      756.632454  |
|  4096.0  |   7168.0  |  4096.0  | 8.0 |   2.0 |   1.319385 |    375.762263   |      471.566273  |
|    64.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   0.433888 |     34.459348   |     2176.419178  |
|   128.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   0.417260 |     70.729104   |     2250.802927  |
|   256.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   0.489365 |    120.693798   |     1971.764433  |
|   512.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   0.571987 |    232.822510   |     1675.010347  |
|  1024.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   0.960137 |    265.267443   |     1279.416592  |
|  2048.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   1.366209 |    368.979210   |      802.725164  |
|  4096.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   2.373175 |    418.532396   |      523.953980  |

EVEN_K enabled performance (model):
|       Model      |    M    |      N    |     K    |  E  | top_k | Time (ms) |      TFLOPS     | Bandwidth (GB/s) |
|------------------|---------|-----------|----------|-----|-------|-----------|-----------------|------------------|
|   mistral-7B     |   4096  |   14336   |   4096   |  8  |   2   |   2.293706 |     405.064608  |      531.122608  |
|   mistral-7B     |   4096  |    4096   |   7168   |  8  |   2   |   1.220967 |     394.339966  |      492.931517  |
|  mistral-22B     |   4096  |   16384   |   6144   |  8  |   2   |   4.106346 |     397.809073  |      453.992305  |
|  mistral-22B     |   4096  |    6144   |   8192   |  8  |   2   |   2.014467 |     412.967996  |      490.271498  |

original performance:
|     M    |      N    |     K    |  E  | top_k | Time (ms) |      TFLOPS     | Bandwidth (GB/s) |
|----------|-----------|----------|-----|-------|-----------|-----------------|------------------|
|    64.0  |    256.0  |   128.0  | 8.0 |   2.0 |   0.081925 |      0.105903   |        7.687693  |
|    64.0  |   1792.0  |  1024.0  | 8.0 |   2.0 |   0.090142 |      5.449960   |      355.150773  |
|    64.0  |   7168.0  |  4096.0  | 8.0 |   2.0 |   0.262167 |     28.660762   |     1795.403410  |
|   128.0  |   7168.0  |  4096.0  | 8.0 |   2.0 |   0.267455 |     56.110850   |     1771.925168  |
|  1024.0  |   7168.0  |  4096.0  | 8.0 |   2.0 |   0.651711 |    182.734773   |      779.381701  |
|  4096.0  |   7168.0  |  4096.0  | 8.0 |   2.0 |   1.288135 |    369.836013   |      482.546716  |
|    64.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   0.437519 |     34.283575   |     2165.703901  |
|   128.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   0.427997 |     70.857381   |     2210.364209  |
|   256.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   0.495296 |    122.781036   |     2025.930717  |
|   512.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   0.585848 |    199.926815   |     1593.819785  |
|  1024.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   1.023181 |    246.732024   |     1045.215086  |
|  2048.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   1.364827 |    362.329243   |      792.246487  |
|  4096.0  |  14336.0  |  4096.0  | 8.0 |   2.0 |   2.361838 |    413.420601   |      520.886676  |

original performance (model):
|       Model      |    M    |      N    |     K    |  E  | top_k | Time (ms) |      TFLOPS     | Bandwidth (GB/s) |
|------------------|---------|-----------|----------|-----|-------|-----------|-----------------|------------------|
|   mistral-7B     |   4096  |   14336   |   4096   |  8  |   2   |   2.300467 |     416.588663  |      524.333165  |
|   mistral-7B     |   4096  |    4096   |   7168   |  8  |   2   |   1.214763 |     395.404826  |      488.501595  |
|  mistral-22B     |   4096  |   16384   |   6144   |  8  |   2   |   3.948421 |     394.402199  |      459.016435  |
|  mistral-22B     |   4096  |    6144   |   8192   |  8  |   2   |   2.000019 |     411.912343  |      484.526327  |

<!---
The core Triton is a small number of people, and we receive many PRs (thank
you!).  To help us review your code more quickly, **if you are a new
contributor (less than 3 PRs merged) we ask that you complete the following
tasks and include the filled-out checklist in your PR description.**

Complete the following tasks before sending your PR, and replace `[ ]` with
`[x]` to indicate you have done them.
-->

# New contributor declaration
- [X] I am not making a trivial change, such as fixing a typo in a comment.

- [X] I have written a PR description following these
  [rules](https://cbea.ms/git-commit/#why-not-how).

- [X] I have run `pre-commit run --from-ref origin/main --to-ref HEAD`.

- Select one of the following.
  - [ ] I have added tests.
    - `/test` for `lit` tests
    - `/unittest` for C++ tests
    - `/python/test` for end-to-end tests
  - [X] This PR does not need a test because `its performance tuning`.

- Select one of the following.
  - [ ] I have not added any `lit` tests.
  - [ ] The `lit` tests I have added follow these [best practices](https://mlir.llvm.org/getting_started/TestingGuide/#filecheck-best-practices),
    including the "tests should be minimal" section. (Usually running Python code
    and using the instructions it generates is not minimal.)